Changed how total addresses are calculated
This only affects when subnet mask bytes are passed in the constructor.
I've switched it from calculating it from the subnet mask's last used
octet, to calculating it from the CIDR notation it already calculated. I
will be returning back to fix this.

This was caused by when the last used octet has 0 unused bits. It would
always return 0 total addresses.